Minor changes to avoid bugs in GCC 4.1.2 on Debian Etch

+++ dlib/lsh/create_random_projection_hash.h
@@ -79,7 +79,7 @@ namespace dlib
         // figure out what the offset values should be
         for (int itr = 0; itr < offset.size(); ++itr)
         {
-            counts.assign(std::pow(2.0,bits), 0);
+            counts.assign(static_cast<unsigned long>(std::pow(2.0,bits)), 0);
             // count the popularity of each hash value
             for (unsigned long i = 0; i < h.size(); ++i)
             {
